requirements.txt
pydicom
Pillow
numpy
pylibjpeg
pylibjpeg-libjpeg
pylibjpeg-openjpeg